Refactor castling rights and some parts of the board, add missing pragma once's
