+++ Sirius/src/castling.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "attacks.h"
 #include "defs.h"
 #include "util/enum_array.h"
@@ -10,6 +12,120 @@ enum class CastleSide
 
 template<typename T>
 using CastleSideArray = EnumArray<T, CastleSide, 2>;
+
+struct CastlingRights
+{
+public:
+    enum class Internal : uint8_t
+    {
+        NONE = 0,
+        WHITE_KING_SIDE = 1,
+        WHITE_QUEEN_SIDE = 2,
+        BLACK_KING_SIDE = 4,
+        BLACK_QUEEN_SIDE = 8,
+        ALL = 15
+    };
+    static constexpr Internal NONE = Internal::NONE;
+    static constexpr Internal WHITE_KING_SIDE = Internal::WHITE_KING_SIDE;
+    static constexpr Internal WHITE_QUEEN_SIDE = Internal::WHITE_QUEEN_SIDE;
+    static constexpr Internal BLACK_KING_SIDE = Internal::BLACK_KING_SIDE;
+    static constexpr Internal BLACK_QUEEN_SIDE = Internal::BLACK_QUEEN_SIDE;
+    static constexpr Internal ALL = Internal::ALL;
+
+    constexpr CastlingRights();
+    constexpr CastlingRights(Internal v);
+    constexpr CastlingRights(Color color, CastleSide side);
+
+    constexpr CastlingRights& operator&=(const CastlingRights& other);
+    constexpr CastlingRights& operator|=(const CastlingRights& other);
+
+    constexpr CastlingRights operator&(const CastlingRights& other) const;
+    constexpr CastlingRights operator|(const CastlingRights& other) const;
+
+    constexpr CastlingRights operator~() const;
+
+    constexpr bool has(CastlingRights other) const;
+    constexpr bool has(Internal v) const;
+
+    constexpr int value() const;
+
+private:
+    Internal m_Value;
+};
+
+constexpr CastlingRights operator&(CastlingRights::Internal a, CastlingRights::Internal b)
+{
+    return CastlingRights(
+        static_cast<CastlingRights::Internal>(static_cast<int>(a) & static_cast<int>(b)));
+}
+
+constexpr CastlingRights operator|(CastlingRights::Internal a, CastlingRights::Internal b)
+{
+    return CastlingRights(
+        static_cast<CastlingRights::Internal>(static_cast<int>(a) | static_cast<int>(b)));
+}
+
+constexpr CastlingRights operator~(CastlingRights::Internal a)
+{
+    return static_cast<CastlingRights::Internal>(~static_cast<int>(a)) & CastlingRights::ALL;
+}
+
+constexpr CastlingRights::CastlingRights()
+    : m_Value(Internal::NONE)
+{
+}
+
+constexpr CastlingRights::CastlingRights(Internal v)
+    : m_Value(v)
+{
+}
+
+constexpr CastlingRights::CastlingRights(Color color, CastleSide side)
+    : m_Value(static_cast<Internal>(1 << (2 * static_cast<int>(color) + static_cast<int>(side))))
+{
+}
+
+constexpr CastlingRights& CastlingRights::operator&=(const CastlingRights& other)
+{
+    *this = *this & other;
+    return *this;
+}
+
+constexpr CastlingRights& CastlingRights::operator|=(const CastlingRights& other)
+{
+    *this = *this | other;
+    return *this;
+}
+
+constexpr CastlingRights CastlingRights::operator&(const CastlingRights& other) const
+{
+    return m_Value & other.m_Value;
+}
+
+constexpr CastlingRights CastlingRights::operator|(const CastlingRights& other) const
+{
+    return m_Value | other.m_Value;
+}
+
+constexpr CastlingRights CastlingRights::operator~() const
+{
+    return ~m_Value;
+}
+
+constexpr bool CastlingRights::has(CastlingRights other) const
+{
+    return static_cast<int>((m_Value & other.m_Value).m_Value) != 0;
+}
+
+constexpr bool CastlingRights::has(Internal v) const
+{
+    return static_cast<int>((m_Value & v).m_Value) != 0;
+}
+
+constexpr int CastlingRights::value() const
+{
+    return static_cast<int>(m_Value);
+}
 
 struct CastlingData
 {
