+++ Sirius/src/cuckoo.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "defs.h"
 #include <array>
