+++ Sirius/src/misc.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "board.h"
 #include "movegen.h"
