+++ Sirius/src/util/string_split.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <sstream>
 #include <string>
 #include <string_view>
